chore&refactor: adjust project structure and update readme
