+++ pkg/parser/newINI.txt
@@ -1,0 +1,7 @@
+[database]
+file=payroll.dat
+port=143
+server=192.0.2.62
+[owner]
+name=WalterWhite
+organization=AcmeWidgetsInc.
